cloud-functions/broadband-scraper/requirements.txt
functions-framework==3.*
requests==2.31.0
google-generativeai==0.3.2
